+++ ios/App/App/LocalFileServerPlugin.m
@@ -1,7 +1,0 @@
-#import <Capacitor/Capacitor.h>
-
-CAP_PLUGIN(LocalFileServerPlugin, "LocalFileServer",
-    CAP_PLUGIN_METHOD(getServerUrl, CAPPluginReturnPromise);
-    CAP_PLUGIN_METHOD(startFileServer, CAPPluginReturnPromise);
-    CAP_PLUGIN_METHOD(refreshFolderAccess, CAPPluginReturnPromise);
-)